requirements.txt
discord.py == 1.4.1
python-dotenv
asyncio
psutil
traceback2
six
rsa